moves files into labelled subdirectories
